requirements.txt
azure-keyvault
azure-mgmt-keyvault==0.30.1